+++ common/src/main/java/com/wynntils/models/character/event/CharacterDeathEvent.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.characterstats.event;
+package com.wynntils.models.character.event;
 
 import com.wynntils.utils.mc.type.Location;
 import net.minecraftforge.eventbus.api.Event;
